Redirect Incoming Requests
If there is a setting for the RedirectUrl, then in the
Application_BeginRequest the Response will be redirected to the value
set in the configuration value.

CloudConfigurationManager has been used instead of ConfigurationManager
incase this comes into the ES5 branch. Default behaviour is to fallback
to Web.config setting if no setting exists in the cloud configuration